Refactorings needed to share the generate_broker_key logic with Origin
The main difference between the logic in Origin was the use of the :rhlogin key
in the encrypted token.  We can't change that because tokens in the wild will
still have :rhlogin in them.

Note: Apart from the automated tests I spun up a devenv without my change and
generated a broker iv/key pair.  I then passed that in to a devenv that had
this code and it was able to decrypt and parse everything correctly.

The other difference was the fact that check_broker_key became
validate_broker_key in Origin.  It's important to know that the method
signature for this method in crankcase is different from check_broker_key.  I
have no idea why this was done originally but it's critical to know that it's
now validate_broker_key(iv, key) instead of check_broker_key(key, iv).

I did talk with Thomas Weist about these configuration changes.  He was OK with
them shipping with 2.0.18.