Make enterprise banner responsive
I added styling to make the enterprise banner responsive. It's in a
separate sub-stylesheet _enterprise.scss, so that it should be easy to
remove when we no longer need the banner. I also added content to the
banner in order to more easily make it responsive. I made it a
sub-section and put the content in the home page. This means it will
only render on the home page, and this should also make it easier to
remove.